Flytter mapping av K9Format til riktig klasse
